requirements.txt
pyserial==3.1.1
whiptail==0.2
